+++ src/mulle-bootstrap-project.sh
@@ -30,6 +30,7 @@
 #
 MULLE_BOOTSTRAP_PROJECT_SH="included"
 
+
 project_usage()
 {
    cat <<EOF >&2